Finished Backend and Frontend

+++ Backend/routes/HandlePools.js
@@ -1,6 +1,7 @@
 const express = require("express");
 const router = express.Router();
 const Pool = require("../pool");
+const jwt = require('jsonwebtoken');
 
 
 router.get('/', async (req, res) => {
@@ -12,16 +13,12 @@ router.get('/', async (req, res) => {
     }
 })
 
-router.get('/:id', getPool, (req, res) => {
-    res.send(res.Pool)
-})
-
-router.post('/', async (req, res) => {
+router.post('/', authenticateToken, async (req, res) => {
     const pool = new Pool({
         Title: req.body.Title,
         Options: req.body.Options,
         MultipleChoice: req.body.MultipleChoice,
-        Owner: req.body.Owner
+        Owner: req.user._id
     })
 
     try {
@@ -32,30 +29,58 @@ router.post('/', async (req, res) => {
     }
 })
 
-router.patch('/:id', getPool, async (req, res) => {
+router.patch('/:id', [authenticateToken, getPool], async (req, res) => {
+    const isInArray = res.Pool.Voters.some((Voter) => {
+        return Voter.equals(req.user._id);
+    });
 
-    for(let i = 0; i < 3; i++) {
-        res.Pool.Options[i].votes += req.body.Options[i];
+    if (isInArray) {
+        res.status(403).json({ Message: "User already voted" })
     }
+    else {
+        res.Pool.Voters.push(req.user._id)
 
-    try{
-        const modPool = await res.Pool.save()
-        res.status(200).json(modPool)
-    } catch (err) {
-        res.status(500).json({ Message: err.message })
+        for (let i = 0; i < 3; i++) {
+            res.Pool.Options[i].votes += req.body.Options[i];
+        }
+
+        try {
+            const modPool = await res.Pool.save()
+            res.status(200).json(modPool)
+        } catch (err) {
+            res.status(500).json({ Message: err.message })
+        }
     }
 })
 
-router.delete('/:id', async (req, res) => {
-    try{
-        await Pool.findByIdAndDelete(req.params.id)
-        res.status(200).json({ Message: "Pool deleted" })
+router.delete('/:id', [authenticateToken, getPool], async (req, res) => {
+    try {
+        if (req.user._id == res.Pool.Owner) {
+            await Pool.findByIdAndDelete(req.params.id)
+            res.status(200).json({ Message: "Pool deleted" })
+        }
+        else {
+            res.status(403).json({ Message: "Not the owner" })
+        }
     } catch (err) {
         res.status(500).json({ Message: err.message })
     }
 
 
 })
+
+function authenticateToken(req, res, next) {
+    const authHeader = req.headers['authorization']
+    const token = authHeader && authHeader.split(' ')[1]
+
+    if (token == null) return res.sendStatus(401)
+
+    jwt.verify(token, process.env.ACCESS_TOKEN_SECRET, (err, user) => {
+        if (err) return res.sendStatus(403)
+        req.user = user
+        next()
+    })
+}
 
 async function getPool(req, res, next) {
     let pool;
@@ -65,7 +90,7 @@ async function getPool(req, res, next) {
         if (pool == null) {
             return res.status(404).json({ Message: "Cannot find pool" })
         }
-    } catch(err) {  
+    } catch (err) {
         return res.status(500).json({ Message: err.message })
     }
